Ensure energy sample count is always an integer.

+++ nion/eels_analysis/EELS_CrossSections.py
@@ -31,7 +31,7 @@ def k_shell_hydrogenic_gos(atomic_number: int, edge_onset_eV: float, edge_delta_
     beamGamma = 1 + beam_energy_eV / electronRestEnergy_eV
     beamBeta2 = 1 - 1 / beamGamma ** 2
 
-    energySampleCount = numpy.fmin(numpy.fmax(50, 100 * numpy.round(edge_delta_eV / 100)), 1000) + 1
+    energySampleCount = int(numpy.fmin(numpy.fmax(50, 100 * numpy.round(edge_delta_eV / 100)), 1000) + 1)
     thetaSampleCount = int(numpy.fmin(numpy.fmax(50, 100 * numpy.round(collection_angle_rad / 0.050)), 400) + 1)
 
     screenedZ2 = 1.
